+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Changed Categories & Brands validation to One of Range
+
 ## [0.2.0] - 2021-11-08
 
 ### Added